# doc/api_docs/openapi.bundled.yaml
openapi: 3.1.0
info:
  title: Treasure Doc API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    Treasure Doc（个人博客 + 后台管理）的 HTTP API 规格。

    ## 文件组织
    本仓库的 API 规格采用「**一个端点一个文件**」的组织方式：

    - `openapi.yaml`（本文件）：聚合根。只声明 `info` / `servers` / `tags` / `security` 与全部 `paths`，
      每个 operation（`get` / `post` / `put` / `patch` / `delete`）通过 $ref 指向对应端点文件。
    - `blog/`（22 个）：公开只读博客 API `/api/blog/*` 与站点根 `robots.txt` / `sitemap.xml` / `rss.xml`。
    - `blog-mgr/`（80 个）：后台管理 API `/api/blog-mgr/*`（含 `/resources/<资源>/` 子目录、备份、媒体库、编辑历史）。
    - `memo/`（7 个）：前台速记本 API `/api/memo/*`（登录即博主本人）。
    - `publish/`（12 个）：机器令牌发布 API `/api/publish/*`。
    - `user/`（5 个）：鉴权 API `/api/user/*` 与健康检查 `/ping`。
    - `backup/`（1 个）：NAS 机器令牌导出 `/api/backup/export`。
    - `components/`：跨端点共享的 `schemas`、`responses`、`parameters`，端点文件用相对路径引用。

    端点文件的内容是一个 **operation 对象**（不含 `openapi:` / `paths:` 包装），首行注释写明
    「HTTP 方法 + 完整路径 + 所属模块 + 路由注册位置」。

    ## 解析方式
    这是「subdocuments」风格的多文件规格，**不能**被 Swagger UI 直接加载，需先打包：

    ```bash
    npx @redocly/cli bundle doc/api_docs/openapi.yaml -o /tmp/openapi.bundled.yaml
    ```

    然后即可用 Swagger UI / Redoc / `npx @redocly/cli preview-docs doc/api_docs/openapi.yaml` 浏览。

    ## 统一响应约定
    - 除 `GET /ping` 与站点根三个文件外，所有接口都返回 `{code, msg, data}` 信封。
    - 业务成功与失败通常都返回 HTTP 200，靠 `code` 区分（`0` = 成功）；
      认证/权限/限流中间件则返回真实的 401 / 403 / 429。
...

    ## 鉴权
    | 请求头 | 适用接口 | 说明 |
    | --- | --- | --- |
...

...
servers:
...
security:
  - {}
...
paths:
...
  /api/memo/memos:
    get:
      operationId: listMemos
      tags:
        - memo-manage
      summary: 我的速记列表（分页，含私有）
      description: |
        返回博主本人的全部速记（含私有与公开）。排序为 `pinned DESC` + `updated_at DESC`；
        `keyword` 模糊匹配标题与正文。登录即博主本人，不要求管理员权限。
      security:
        - XToken: []
      parameters:
        - $ref: '#/components/parameters/ManagePage'
        - $ref: '#/components/parameters/ManagePageSize'
        - $ref: '#/components/parameters/ManageKeyword'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemoPage'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '500':
          $ref: '#/components/responses/ManageInternal'
    post:
      operationId: createMemo
      tags:
        - memo-manage
      summary: 新建速记
      description: |
        新建一条速记。`public` 缺省为 false（仅自己可见）；新建不校验 `version`。
      security:
        - XToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemoRecord'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '500':
          $ref: '#/components/responses/ManageInternal'
  /api/memo/memos/batch-delete:
    post:
      operationId: deleteMemos
      tags:
        - memo-manage
      summary: 批量删除速记（软删除）
      description: |
        按 ID 批量软删除速记（`deleted_at` 置位）。ID 会去重，空集合或超过 200 个返回 400 / code 40001。
      security:
        - XToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageBatchDeleteInput'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ManageBatchDeleteResult'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '500':
          $ref: '#/components/responses/ManageInternal'
  /api/memo/memos/{id}:
    get:
      operationId: getMemo
      tags:
        - memo-manage
      summary: 单条速记详情
      description: |
        返回单条速记（含正文与私有信息）。仅博主本人可访问。
      security:
        - XToken: []
      parameters:
        - $ref: '#/components/parameters/ManageID'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemoRecord'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '404':
          $ref: '#/components/responses/ManageNotFound'
        '500':
          $ref: '#/components/responses/ManageInternal'
    put:
      operationId: updateMemo
      tags:
        - memo-manage
      summary: 更新速记
      description: |
        更新单条速记，必须携带当前 `version`（乐观锁）；版本过期返回 409 / code 40900。
        更新会维护 `publicAt`：首次公开时写入当前时间，取消公开则清零。
      security:
        - XToken: []
      parameters:
        - $ref: '#/components/parameters/ManageID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemoRecord'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '404':
          $ref: '#/components/responses/ManageNotFound'
        '409':
          $ref: '#/components/responses/ManageConflict'
        '500':
          $ref: '#/components/responses/ManageInternal'
    delete:
      operationId: deleteMemo
      tags:
        - memo-manage
      summary: 删除速记（软删除）
      description: |
        删除为 GORM 软删除（`deleted_at` 置位），不提供前端恢复入口。
      security:
        - XToken: []
      parameters:
        - $ref: '#/components/parameters/ManageID'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ManageDeleteResult'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '404':
          $ref: '#/components/responses/ManageNotFound'
        '500':
          $ref: '#/components/responses/ManageInternal'
  /api/memo/memos/{id}/fields:
    patch:
      operationId: patchMemoFields
      tags:
        - memo-manage
      summary: 快捷修改速记字段
      description: |
        黑名单外的单个字段快捷修改，白名单：`pinned` / `public`。不要求 `version`，成功后自动自增。
        修改 `public` 时同样维护 `publicAt`。
      security:
        - XToken: []
      parameters:
        - $ref: '#/components/parameters/ManageID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoUpdateFieldsInput'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/MemoRecord'
        '400':
          $ref: '#/components/responses/ManageInvalidRequest'
        '401':
          $ref: '#/components/responses/ManageUnauthorized'
        '404':
          $ref: '#/components/responses/ManageNotFound'
        '500':
          $ref: '#/components/responses/ManageInternal'
...
components:
  securitySchemes:
    XToken:
      type: apiKey
      in: header
      name: X-Token
      description: 登录成功后签发的用户令牌；后台接口还要求 `userType` ∈ {2, 100}。
    XPublishToken:
      type: apiKey
      in: header
      name: X-Publish-Token
      description: 机器令牌；取值为后台「站点设置 → 发布令牌」配置的值，未配置时所有 `/api/publish/*` 请求返回 401。
    XBackupToken:
      type: apiKey
      in: header
      name: X-Backup-Token
      description: NAS 机器令牌；取值为后台配置的备份令牌（回退配置文件 `[backup].apiToken`）。
  schemas:
    Envelope:
      type: object
      description: |
        统一响应信封，所有接口（除 GET /ping 与 robots/sitemap/rss 外）都返回该结构。
        业务结果由 code 表达：0 = 成功，其余为各模块业务码。失败时 data 通常为 null，
        字段级校验失败时 data 为 `{"field": "..."}`（管理端）或 `{}`（用户模块）。
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务码，0 表示成功
          example: 0
        msg:
          type: string
          description: 提示信息；成功时为空字符串
          example: ''
        data:
          description: |
            业务数据；失败时通常为 null。各端点在此声明具体结构
            （对象、数组或 null）；未声明的端点表示任意值。
            此处不限定类型，以便端点用 allOf 追加自己的 data 结构。
...
    Pagination:
      type: object
      description: 分页信息
      required:
        - page
        - pageSize
        - total
        - orderBy
      properties:
        page:
          type: integer
          description: 当前页码，从 1 开始
          example: 1
        pageSize:
          type: integer
          description: 每页条数
          example: 20
        total:
          type: integer
          format: int64
          description: 满足条件的总条数
          example: 42
        orderBy:
          type: string
          description: 实际生效的排序表达式（服务端白名单校验后生成，非请求原值）
          example: created_at desc
...
    ManageBatchDeleteInput:
      type: object
      description: 批量软删除载荷；ID 会去重，空集合或超过 200 个返回 400 / code 40001
      required:
        - ids
      properties:
        ids:
          type: array
          maxItems: 200
          items:
            type: string
      example:
        ids:
          - '1930526412345678901'
          - '1930526412345678902'
    ManageBatchDeleteResult:
      type: object
      description: 批量软删除结果
      required:
        - deleted
      properties:
        deleted:
          type: integer
          format: int64
          description: 实际删除条数
          example: 2
    ManageDeleteResult:
      type: object
      description: 单条删除结果（GORM 软删除）
      required:
        - deleted
      properties:
        deleted:
          type: boolean
          example: true
...
    MemoRecord:
      type: object
      description: 速记记录（博主管理端返回，含正文、图片与私有信息）
      required:
        - id
        - title
        - content
        - images
        - pinned
        - public
        - tags
        - sortOrder
        - version
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: 雪花 ID
          example: '1930526412345678901'
        title:
          type: string
          description: 速记标题，可留空（列表用正文首句代替）
          example: 灵感笔记
        content:
          type: string
          description: 正文（纯文本，非 Markdown）
        images:
          type: array
          items:
            type: string
          description: 图片路径数组（`/files/blog/...`）；保证为空数组而非 null
        pinned:
          type: boolean
          description: 是否置顶
        public:
          type: boolean
          description: 是否对访客公开（默认私有）
        tags:
          type: array
          items:
            type: string
          description: 标签名数组；保证为空数组而非 null
        sortOrder:
          type: integer
          description: 自定义排序权值（预留，暂未参与排序）
        version:
          type: integer
          description: 乐观锁版本；更新时必填且必须等于当前版本
        createdAt:
          type: string
          format: date-time
          description: 创建时间（RFC3339）
        updatedAt:
          type: string
          format: date-time
          description: 最后更新时间（RFC3339）
    MemoPage:
      type: object
      description: 速记管理列表分页结果
      required:
        - list
        - pagination
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/MemoRecord'
        pagination:
          $ref: '#/components/schemas/Pagination'
    MemoInput:
      type: object
      description: 速记写入载荷（新建与更新共用；新建默认私有 public=false，更新需带 version）。正文必填（不支持 Markdown）。
      required:
        - content
      properties:
        title:
          type: string
          description: 速记标题，可留空
        content:
          type: string
          description: 正文（纯文本），不能为空
        images:
          type: array
          items:
            type: string
          description: 图片路径数组（`/files/blog/...`）
        pinned:
          type: boolean
          default: false
        public:
          type: boolean
          default: false
          description: 是否对访客公开；受站点总开关 `memoPublicEnabled` 约束
        tags:
          type: array
          items:
            type: string
          description: 标签名数组
        sortOrder:
          type: integer
          default: 0
        version:
          type: integer
          description: 乐观锁版本；更新时必填且必须等于当前版本，否则 409 / code 40900
    MemoUpdateFieldsInput:
      type: object
      description: 速记快捷修改载荷，白名单 `pinned` / `public`；成功后服务端自增 version
      properties:
        pinned:
          type: boolean
        public:
          type: boolean
      example:
        pinned: true
  responses:
...
    ManageInvalidRequest:
      description: '请求参数格式错误（HTTP 400，code 40001）；字段级校验失败时 msg 为具体原因，且 data 为 {"field": "<字段名>"}'
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type:
                      - object
                      - 'null'
                    description: '参数格式错误时为 null；字段级校验失败时为 {"field": "<字段名>"}'
                    properties:
                      field:
                        type: string
                        description: 出错的 JSON 字段名，仅字段级校验失败时存在
          examples:
            invalid:
              value:
                code: 40001
                msg: 请求参数格式错误
                data: null
            fieldError:
              value:
                code: 40001
                msg: 名称不能为空
                data:
                  field: name
    ManageUnauthorized:
      description: 未登录或登录已失效（HTTP 401，code 40100）；`X-Token` 缺失或无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 40100
            msg: 未登录或登录已失效
            data: null
...
    ManageInternal:
      description: 服务内部错误（HTTP 500，code 50000）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 50000
            msg: 服务内部错误
            data: null
    ManageConflict:
      description: 版本冲突或标识重复（HTTP 409，code 40900）；`version` 过期或 slug / 标签名等唯一标识已存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 40900
            msg: 数据已变更或标识重复
            data: null
    ManageNotFound:
      description: 资源不存在（HTTP 404，code 40410）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Envelope'
          example:
            code: 40410
            msg: 资源不存在
            data: null
...
  parameters:
...
    ManagePage:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始
      schema:
        type: integer
        minimum: 1
        default: 1
    ManagePageSize:
      name: pageSize
      in: query
      required: false
      description: 每页条数
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ManageKeyword:
      name: keyword
      in: query
      required: false
      description: 关键词模糊搜索
      schema:
        type: string
...
    ManageID:
      name: id
      in: path
      required: true
      description: 资源 ID（雪花 ID）
      schema:
        type: string
        minLength: 1
        maxLength: 128
...